requirements.txt
Django>=5.0
psycopg2-binary>=2.9
djangorestframework>=3.14.0
django-filter>=23.2
